chore(connlib): mitigate WARN logs from phoenix-channel (#6759)
Merging #6708 had an unintended side-effect that we are seeing a lot of
WARN logs from phoenix-channel because we can no longer parse the
response from gateways. We didn't do anything with these responses but
gateways are sending them for backwards-compatibility reasons.

To not confuse ourselves while debugging, we revert the client-side bit
of #6708 to remove these warnings.