fix(connlib): blackhole DNS HTTPS type queries for resources (#6788)
Fix #6781
Fix #6375

The problem was that browsers in iOS(and possible other OSes) queries
for A, AAAA and HTTPS, and we correctly intercept A and AAAA.

Correctly intercepting HTTPS queries is more tricky since we need the
server's alpn, before this PR we were just forwarding those and then the
response back but the problem with that is that it'd return the real IP
for the service instead of our proxy IP.

So to quickly fix this we simply blackhole the query so the browser
never use that response.

In the future an improvement over this would be to intercept the
response instead of the query and mangle the ips there.

---------

Signed-off-by: Gabi <gabrielalejandro7@gmail.com>
Co-authored-by: Reactor Scram <ReactorScram@users.noreply.github.com>