chore(snownet): unset parent span (#6865)
When constructing a span, any currently set span will automatically be
set as the parent. In the case of the `connection` span, this was the
`accept_answer` or `new_connection` span from the client / gateway.
Those are not meant to be re-activated every time we enter the
`connection` span.

By setting an explicit parent, we avoid that.

Unfortunately, this means that this span will never have a parent, even
if other spans are active whilst we enter this one. We enter this one in
the hot-path, which is why it is being constructed ahead of time.

---------

Signed-off-by: Thomas Eizinger <thomas@eizinger.io>