refactor(snownet): only initiate a single WireGuard session (#6936)
To reduce the TTFB, we immediately force a WireGuard handshake as soon
as ICE completes. Currently, both the client and the gateway do this
which is somewhat counter-productive as there can only be one active
session.

With this patch, only the client will force a new WireGuard handshake as
soon as ICE completes.