fix(connlib): default to 443 for websocket endpoint (#6961)
In #6909, we introduced a regression that wasn't caught by CI.
Previously, we were using a different function to resolve the domain
name of the portal. That function took care of handling the case where
the host didn't have a port number.

In the docker-compose file we always specify a port number, therefore
the case of host-only doesn't get tested.

This currently prevents all clients from signing in to staging & prod.